This is my initial commit.
I copied this from the mtiller/ModelicaTest repository.  I moved it to the xogeny user, since it really belongs there and I renamed it to XogenyTest because a) ModelicaTest was already taken and b) I wanted a name that more accurately reflected that this was not an official Modelica library in any way.